requirements.txt
Django==5.0.3
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
reportlab==4.1.0
python-dateutil==2.9.0
django-import-export==3.3.6
djangorestframework==3.15.0
Pillow==10.2.0
XlsxWriter==3.2.0
tablib==3.5.0  # Required by django-import-export 3.3.6
